requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
biopython==1.83
supabase==2.0.3
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1